# api/openapi.json
{
  "components": {
    "schemas": {
...
      "AiChatMessage": {
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "AiChatMessage",
        "type": "object"
      },
      "AiChatRequest": {
        "properties": {
          "birth_details": {
            "$ref": "#/components/schemas/BirthDetails"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/AiChatMessage"
            },
            "maxItems": 11,
            "minItems": 1,
            "title": "Messages",
            "type": "array"
          }
        },
        "required": [
          "birth_details",
          "messages"
        ],
        "title": "AiChatRequest",
        "type": "object"
      },
      "AiChatResponse": {
        "properties": {
          "questions_remaining": {
            "title": "Questions Remaining",
            "type": "integer"
          },
          "questions_used": {
            "title": "Questions Used",
            "type": "integer"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          }
        },
        "required": [
          "reply",
          "questions_used",
          "questions_remaining"
        ],
        "title": "AiChatResponse",
        "type": "object"
      },
...
      "AstrologerOnboardingRequest": {
        "properties": {
          "astrology_types": {
            "items": {
              "type": "string"
            },
            "title": "Astrology Types",
            "type": "array"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "experience_years": {
            "title": "Experience Years",
            "type": "integer"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id_proof_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Proof Url"
          },
          "languages": {
            "title": "Languages",
            "type": "string"
          },
          "legal_agreement_accepted": {
            "default": true,
            "title": "Legal Agreement Accepted",
            "type": "boolean"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "phone_number": {
            "title": "Phone Number",
            "type": "string"
          },
          "preferred_working_hours": {
            "title": "Preferred Working Hours",
            "type": "string"
          },
          "profile_photo_url": {
            "title": "Profile Photo Url",
            "type": "string"
          },
          "short_bio": {
            "title": "Short Bio",
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "email",
          "phone_number",
          "password",
          "astrology_types",
          "experience_years",
          "languages",
          "preferred_working_hours",
          "short_bio",
          "profile_photo_url"
        ],
        "title": "AstrologerOnboardingRequest",
        "type": "object"
      },
...
      "BirthDetails": {
        "properties": {
          "date_of_birth": {
            "format": "date",
            "title": "Date Of Birth",
            "type": "string"
          },
          "gender": {
            "$ref": "#/components/schemas/GenderType"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "place_of_birth": {
            "maxLength": 150,
            "minLength": 1,
            "title": "Place Of Birth",
            "type": "string"
          },
          "time_of_birth": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Of Birth"
          }
        },
        "required": [
          "name",
          "date_of_birth",
          "place_of_birth",
          "gender"
        ],
        "title": "BirthDetails",
        "type": "object"
      },
      "Body_login_login_post": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          },
          "grant_type": {
            "anyOf": [
              {
                "pattern": "^password$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_login_post",
        "type": "object"
      },
...
      "Body_upload_onboarding_photo_astrologers_onboarding_photo_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_onboarding_photo_astrologers_onboarding_photo_post",
        "type": "object"
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "type": "object"
      },
      "GenderType": {
        "enum": [
          "MALE",
          "FEMALE",
          "OTHER"
        ],
        "title": "GenderType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResendVerificationRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ResendVerificationRequest",
        "type": "object"
      },
      "ResetPasswordRequest": {
        "properties": {
          "new_password": {
            "title": "New Password",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPasswordRequest",
        "type": "object"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "token_type": {
            "title": "Token Type",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "user_id",
          "role"
        ],
        "title": "Token",
        "type": "object"
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "phone_number": {
            "anyOf": [
              {
                "pattern": "^\\d{10,15}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "role": {
            "$ref": "#/components/schemas/app__schemas__UserRole"
          }
        },
        "required": [
          "password",
          "role"
        ],
        "title": "UserCreate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyOTPRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "otp": {
            "title": "Otp",
            "type": "string"
          },
          "verification_type": {
            "default": "FORGOT_PASSWORD",
            "title": "Verification Type",
            "type": "string"
          }
        },
        "required": [
          "email",
          "otp"
        ],
        "title": "VerifyOTPRequest",
        "type": "object"
      },
...
      "app__schemas__UserRole": {
        "enum": [
          "ADMIN",
          "SEEKER",
          "ASTROLOGER"
        ],
        "title": "UserRole",
        "type": "string"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Aadikarta API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ai-astrologer/chat": {
      "post": {
        "operationId": "ai_chat_ai_astrologer_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Chat",
        "tags": [
          "AI Astrologer"
        ]
      }
    },
...
    "/astrologers/onboarding": {
      "post": {
        "operationId": "astrologer_onboarding_astrologers_onboarding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AstrologerOnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Astrologer Onboarding",
        "tags": [
          "Astrologers"
        ]
      }
    },
    "/astrologers/onboarding/photo": {
      "post": {
        "description": "Public, unauthenticated upload for the applicant's profile photo during\nastrologer onboarding (applicant has no account/token yet at this point).\nRate-limited and restricted to images only \u2014 separate from the admin-only\n/admin/upload endpoint used elsewhere for documents.",
        "operationId": "upload_onboarding_photo_astrologers_onboarding_photo_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_onboarding_photo_astrologers_onboarding_photo_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Onboarding Photo",
        "tags": [
          "Astrologers"
        ]
      }
    },
...
    "/forgot-password": {
      "post": {
        "operationId": "forgot_password_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forgot Password"
      }
    },
...
    "/login": {
      "post": {
        "operationId": "login_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
...
    "/resend-verification": {
      "post": {
        "operationId": "resend_verification_resend_verification_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resend Verification"
      }
    },
    "/reset-password": {
      "post": {
        "operationId": "reset_password_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Password"
      }
    },
...
    "/signup": {
      "post": {
        "operationId": "signup_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup"
      }
    },
...
    "/verify-email": {
      "post": {
        "operationId": "verify_email_verify_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Email"
      }
    },
    "/verify-otp": {
      "post": {
        "operationId": "verify_otp_verify_otp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Otp"
      }
    },
...
  }
}